backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.29
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
httpx>=0.27.0
websockets>=12.0
pyyaml>=6.0.1

# Test dependencies
pytest>=8.2.0
pytest-asyncio>=0.23.6
httpx>=0.27.0
